Shut down webserver when closing connection to sim LCD
Previously, the web server thread not stopping prevented the whole
program (e.g. simple-program.py) from exiting.

Closing the webserver in the closeSerial seems to be the most logical,
as we can expect this method to be called already when the program is
shutting down (as it's the case in simple-program.py for example)

Maybe updating the wiki page about the simulated LCD to mention the need
to call closeSerial to stop the server would be nice too.